fix: eventqueue query cleanup and test fix
